+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Fixed
+
+- Fixed an issue in the `applyNode` GraphQL API where agents could not be
+  properly identified.
+
 ## [0.21.0] - 2024-09-05
 
 ### Added
@@ -650,6 +657,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - An initial version.
 
+[Unreleased]: https://github.com/aicers/review-web/compare/0.21.0...main
 [0.21.0]: https://github.com/aicers/review-web/compare/0.20.0...0.21.0
 [0.20.0]: https://github.com/aicers/review-web/compare/0.19.0...0.20.0
 [0.19.0]: https://github.com/aicers/review-web/compare/0.18.0...0.19.0